+++ src/Helpers/UrlBuilder.php
@@ -89,6 +89,7 @@ class UrlBuilder
     const RESOURCE_GET_USER_SETTINGS = 'resource/UserSettingsResource/get/';
     const RESOURCE_CHANGE_PASSWORD = 'resource/UserPasswordResource/update/';
     const RESOURCE_DEACTIVATE_ACCOUNT = 'resource/DeactivateAccountResource/create/';
+    const RESOURCE_BLOCK_USER = 'resource/UserBlockResource/create/';
 
     /**
      * News